Improve processor

+++ l2awinddirection/l2awinddirection.py
@@ -6,13 +6,37 @@ Step 3: move the generated file to the archive storage
 """
 import glob
 import os
+import pdb
 import shutil
 import xarray as xr
 from tqdm import tqdm
 import logging
 import time
 import numpy as np
+import warnings
 
+warnings.filterwarnings("ignore")
+# logging.getLogger("tensorflow").setLevel(logging.FATAL)
+import tensorflow as tf
+
+tf.get_logger().setLevel("ERROR")
+logging.getLogger("tensorflow").setLevel(logging.ERROR)
+# tf.debugging.set_log_device_placement(True)
+# logging.getLogger("tensorflow").setLevel(logging.WARNING)
+# tf.logging.set_verbosity(tf.logging.ERROR)
+# tf.keras.utils.disable_interactive_logging()
+# import keras
+# from keras.backend.tensorflow_backend import tf
+
+# tf.keras.utils.disable_interactive_logging()
+# import absl.logging
+# logging.root.removeHandler(absl.logging._absl_handler)
+# absl.logging._warn_preinit_stderr = False
+
+# logger = tf.get_logger()
+# logger.setLevel(logging.FATAL)
+
+import l2awinddirection
 from l2awinddirection.generate_L2A_winddir_pdf_product import (
     generate_wind_distribution_product,
 )
@@ -22,6 +46,7 @@ from l2awinddirection.generate_L2A_winddir_regression_product import (
 from l2awinddirection.utils import get_conf
 from l2awinddirection.M64RN4 import M64RN4_distribution, M64RN4_regression
 
+os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"  # disable logging tensorflow
 conf = get_conf()
 
 
@@ -89,6 +114,12 @@ def main():
         required=False,
     )
     parser.add_argument(
+        "--skipmoves",
+        help="skip copy of the tiles in a workspace and move of the L2A winddir files from workspace to outputdir [default=False]",
+        action="store_true",
+        default=False,
+    )
+    parser.add_argument(
         "--remove-tiles",
         help="remove the tiles files in the workspace directory [default=True]",
         required=False,
@@ -133,6 +164,7 @@ def main():
         #     ".../analysis/s1_data_analysis/project_rmarquar/wsat/trained_models/iw/*.hdf5"
         # )
         path_best_models = glob.glob(os.path.join(dirmodelsreg, "*.hdf5"))
+        assert len(path_best_models) > 1
         for path in path_best_models:
 
             m64rn4_reg = M64RN4_regression(input_shape, data_augmentation)
@@ -144,26 +176,43 @@ def main():
     else:
         raise Exception("not handled case")
     # files = glob.glob("/raid/localscratch/agrouaze/tiles_iw_4_wdir/3.1/*SAFE/*.nc")
-    if args.workspace is None:
-        workspace = conf["workspace_prediction"]
+    if args.l2awindirtilessafe.endswith("/"):
+
+        l2awindirtilessafe = args.l2awindirtilessafe.rstrip("/")
     else:
-        workspace = args.workspace
-    logging.info(
-        "workspace where the tiles will be temporarily moved is: %s",
-        workspace,
-    )
-    safefile = os.path.join(workspace, os.path.basename(args.l2awindirtilessafe))
-    if not os.path.exists(safefile):
-        logging.info(" step 1: move %s -> %s", args.l2awindirtilessafe, safefile)
-        shutil.copytree(args.l2awindirtilessafe, safefile)
+        l2awindirtilessafe = args.l2awindirtilessafe
+    if args.skipmoves:
+        workspace = os.path.dirname(l2awindirtilessafe)
+        logging.info(
+            "workspace is the place where the tiles are already stored: %s (no copies)",
+            workspace,
+        )
+    else:
+        if args.workspace is None:
+            workspace = conf["workspace_prediction"]
+        else:
+            workspace = args.workspace
+        logging.info(
+            "workspace where the tiles will be temporarily moved is: %s",
+            workspace,
+        )
+    safefile = os.path.join(workspace, os.path.basename(l2awindirtilessafe))
+    if not os.path.exists(safefile) and args.skipmoves is True:
+        logging.info(" step 1: move %s -> %s", l2awindirtilessafe, safefile)
+        shutil.copytree(l2awindirtilessafe, safefile)
     files = glob.glob(os.path.join(safefile, "*.nc"))
     logging.info("Number of files to process: %d" % len(files))
+    if not os.path.exists(args.outputdir):
+        logging.info("mkdir %s", args.outputdir)
+        os.makedirs(args.outputdir)
+
     logging.info("step 2: predictions")
     for ii in tqdm(range(len(files))):
         file = files[ii]
         #     for file in files:
         tiles = xr.open_dataset(file)
         if len(tiles.variables.keys()) == 0:
+            logging.info("no data in %s", file)
             continue
         if args.mode == "pdf":
             res = generate_wind_distribution_product(
@@ -172,26 +221,37 @@ def main():
             outputfile = file.replace(
                 "_wind.nc", "_winddirection_pdf.nc"
             )  # TODO: change to _tiles.nc -> _winddirection.nc
+            if args.skipmoves:
+                outputfile = outputfile.replace(workspace, args.outputdir)
+                if not os.path.exists(os.path.dirname(outputfile)):
+                    os.makedirs(os.path.dirname(outputfile))
+            res.attrs["winddirection_L2A_processor"] = "l2awinddirection"
+            res.attrs["winddirection_L2A_processor_version"] = l2awinddirection.__version__
             res.to_netcdf(outputfile)
         elif args.mode == "regression":
+
             outputfile = file.replace(
                 "_wind.nc", "_winddirection_regression.nc"
             )  # TODO: change to _tiles.nc -> _winddirection.nc
+            if args.skipmoves:
+                outputfile = outputfile.replace(workspace, args.outputdir)
+                if not os.path.exists(os.path.dirname(outputfile)):
+                    os.makedirs(os.path.dirname(outputfile))
+
             res = generate_wind_product(tiles, model_m64rn4)
+            res.attrs["winddirection_L2A_processor"] = "l2awinddirection"
+            res.attrs["winddirection_L2A_processor_version"] = l2awinddirection.__version__
             res.to_netcdf(outputfile)
+
         # if you want to remove the file containing the tiles used for inference (data kept in final product)
         del tiles
-        if args.remove_tiles:
+        if args.remove_tiles and args.skipmoves is False:
             logging.info("remove temporary tiles file in the workspace: %s", file)
             os.remove(file)
-    final_safe_path = os.path.join(
-        args.outputdir, os.path.basename(args.l2awindirtilessafe)
-    )
-    if not os.path.exists(os.path.dirname(final_safe_path)):
-        logging.info("mkdir %s", os.path.dirname(final_safe_path))
-        os.makedirs(os.path.dirname(final_safe_path))
-    logging.info("step 3: move %s -> %s", safefile, final_safe_path)
-    shutil.move(safefile, final_safe_path)
+    final_safe_path = os.path.join(args.outputdir, os.path.basename(l2awindirtilessafe))
+    if args.skipmoves is False:
+        logging.info("step 3: move %s -> %s", safefile, final_safe_path)
+        shutil.move(safefile, final_safe_path)
     logging.info("Ifremer Level-2A wind direction SAFE path: %s", final_safe_path)
     logging.info("successful SAFE processing")
     logging.info("peak memory usage: %s ", get_memory_usage())
